ci(LH-70586): Ensure release task is run when tag is pushed (#47)
Yesterday, I merged a PR (https://github.com/CiscoDevNet/terraform-provider-cdo/pull/45) where I removed the requirement to run acceptance tests when pushing a tag (having acceptance tests run only on merge to main). However, I failed in that PR to remove the dependency between the `acceptance-tests` job and the `release` job. In this PR, I fix that and add a dependency from unit-test to release instead.