fix(LH-70915): Use external-compute tokens to generate bootstrap data (#55)
* fix(LH-70915): Use external-compute tokens to generate bootstrap data

There is an Anubis weirdness by which only a single token can exist for a given scope (external-compute and fmc tokens are exceptions we deliberately engineered to work around this limitation which is a feature of some kind). So when you call `/oauth/token` with the API token fed in to the Terraform provider, you actually get back the same token you called with. When this token is passed into the bootstrap function of our SDC, the bootstrap method revokes the token after setting up the SDC. This nvalidates the token the terraform-provider is using, and we end up having to regenerate the token. In this PR, I switch to using the external-compute token instead for bootstrapping the SDC. (yes, this is a hack, and no Weilue, we are not going to "fix" this, there have been many discussions around this between Doron and me for months now :) ).

* Actually use the external-compute token